+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased] 
+
+### Added
+
+- [TD-3230] Taxonomy aggregations with enriched information
+
 ## [4.23.0] 2021-06-28
 
 ### Added