+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [TD-3780] Missing `domain_ids` in Audit events 
+- [TD-3971] Template mandatory dependent field
 
 ## [4.27.0] 2021-09-07