+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 - [TD-3780] Missing `domain_ids` in Audit events 
 
+### Fixed
+
+- [TD-4037] change the limit on the taxonomy in aggregations
+
 ## [4.27.0] 2021-09-07
 
 ### Changed